major rework on interrupt manager framework
The generation of the vector table has changed. Now it does not use
preprocessor meta programing at the lower end anymore. Instead, the
whole generation process is fully realised by template meta programing.
All in all the flexibility is increased. As a second issue this patchset
enables the new feature of using a Default Slot within a configuration.

git-svn-id: https://svn-eos.cs.uni-magdeburg.de/repos/Projects/AVR/halib/trunk@755 ac1e7a95-1e17-0410-ad1e-f2b06e490645